requirements.txt
urllib3==1.24.3  # https://github.com/urllib3/urllib3
pytz==2019.1  # https://github.com/stub42/pytz
python-slugify==3.0.4  # https://github.com/un33k/python-slugify
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
psycopg2==2.8  # https://github.com/psycopg/psycopg2
Pygments==2.4.2  # http://pygments.org/
python-dateutil==2.8.0  # https://github.com/dateutil/dateutil
django==3.0.4  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-extensions==2.1.7  # https://github.com/django-extensions/django-extensions
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
flake8==3.7.5  # https://github.com/PyCQA/flake8
pylint==2.3.1  # https://github.com/PyCQA/pylint
pylint-plugin-utils==0.5  # https://github.com/PyCQA/pylint-plugin-utils
pylint-django==2.0.11  # https://github.com/PyCQA/pylint-django
pytest==5.2.1  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.2  # https://github.com/Frozenball/pytest-sugar
pytest-django==3.5.1  # https://github.com/pytest-dev/pytest-django
pytest-mock==1.11.1  # https://github.com/pytest-dev/pytest-mock
pytest-cov==2.8.1  # https://github.com/pytest-dev/pytest-cov
django-coverage-plugin==1.6.0  # https://github.com/nedbat/django_coverage_plugin

# Put your requirement here
django-filter==2.2.0
httpagentparser==1.9.0
redis>=3.2.0
celery==4.4.2

